Rename org.mytake.factset.gui -> .swt, in case we make new guis in the future.
