+++ factset-tooling/src/main/java/org/mytake/factset/swt/video/MismatchCtl.java
@@ -26,7 +26,7 @@
  *
  * You can contact us at team@mytake.org
  */
-package org.mytake.factset.gui.video;
+package org.mytake.factset.swt.video;
 
 
 import com.diffplug.common.base.Either;
@@ -45,8 +45,8 @@ import org.eclipse.swt.widgets.Button;
 import org.eclipse.swt.widgets.Composite;
 import org.eclipse.swt.widgets.Label;
 import org.eclipse.swt.widgets.Text;
-import org.mytake.factset.gui.Labels;
-import org.mytake.factset.gui.Workbench;
+import org.mytake.factset.swt.Labels;
+import org.mytake.factset.swt.Workbench;
 import org.mytake.factset.video.TranscriptMatch;
 import org.mytake.factset.video.Word;
