+++ factset-tooling/src/main/java/org/mytake/factset/swt/video/SaidCtl.java
@@ -26,7 +26,7 @@
  *
  * You can contact us at team@mytake.org
  */
-package org.mytake.factset.gui.video;
+package org.mytake.factset.swt.video;
 
 
 import com.diffplug.common.swt.ControlWrapper;
@@ -40,7 +40,7 @@ import org.eclipse.jface.text.TextViewer;
 import org.eclipse.swt.SWT;
 import org.eclipse.swt.graphics.Point;
 import org.eclipse.swt.widgets.Composite;
-import org.mytake.factset.gui.Labels;
+import org.mytake.factset.swt.Labels;
 import org.mytake.factset.video.SaidTranscript;
 import org.mytake.factset.video.Word;
 import org.mytake.factset.video.Word.Vtt;
