+++ factset-tooling/src/main/java/org/mytake/factset/swt/video/TranscriptCtl.java
@@ -26,7 +26,7 @@
  *
  * You can contact us at team@mytake.org
  */
-package org.mytake.factset.gui.video;
+package org.mytake.factset.swt.video;
 
 
 import com.diffplug.common.io.Files;
@@ -40,8 +40,8 @@ import java.text.SimpleDateFormat;
 import org.eclipse.swt.SWT;
 import org.eclipse.swt.custom.SashForm;
 import org.eclipse.swt.widgets.Composite;
-import org.mytake.factset.gui.Labels;
-import org.mytake.factset.gui.Workbench;
+import org.mytake.factset.swt.Labels;
+import org.mytake.factset.swt.Workbench;
 import org.mytake.factset.video.Ingredients;
 import org.mytake.factset.video.TranscriptMatch;
